+++ src/core/thread/mesh_forwarder.cpp
@@ -40,56 +40,6 @@ namespace ot {
 
 RegisterLogModule("MeshForwarder");
 
-void ThreadLinkInfo::SetFrom(const Mac::RxFrame &aFrame)
-{
-    Clear();
-
-    if (kErrorNone != aFrame.GetSrcPanId(mPanId))
-    {
-        IgnoreError(aFrame.GetDstPanId(mPanId));
-    }
-
-    {
-        Mac::PanId dstPanId;
-
-        if (kErrorNone != aFrame.GetDstPanId(dstPanId))
-        {
-            dstPanId = mPanId;
-        }
-
-        mIsDstPanIdBroadcast = (dstPanId == Mac::kPanIdBroadcast);
-    }
-
-    if (aFrame.GetSecurityEnabled())
-    {
-        uint8_t keyIdMode;
-
-        // MAC Frame Security was already validated at the MAC
-        // layer. As a result, `GetKeyIdMode()` will never return
-        // failure here.
-        IgnoreError(aFrame.GetKeyIdMode(keyIdMode));
-
-        mLinkSecurity = (keyIdMode == Mac::Frame::kKeyIdMode0) || (keyIdMode == Mac::Frame::kKeyIdMode1);
-    }
-    else
-    {
-        mLinkSecurity = false;
-    }
-    mChannel = aFrame.GetChannel();
-    mRss     = aFrame.GetRssi();
-    mLqi     = aFrame.GetLqi();
-#if OPENTHREAD_CONFIG_TIME_SYNC_ENABLE
-    if (aFrame.GetTimeIe() != nullptr)
-    {
-        mNetworkTimeOffset = aFrame.ComputeNetworkTimeOffset();
-        mTimeSyncSeq       = aFrame.ReadTimeSyncSeq();
-    }
-#endif
-#if OPENTHREAD_CONFIG_MULTI_RADIO
-    mRadioType = static_cast<uint8_t>(aFrame.GetRadioType());
-#endif
-}
-
 void MeshForwarder::Counters::UpdateOnTxDone(const Message &aMessage, bool aTxSuccess)
 {
     if (aMessage.GetType() == Message::kTypeIp6)
